user_services/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
djangorestframework==3.13.1
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
urllib3==1.26.9
zope.event==4.5.0
zope.interface==5.4.0
kavenegar==1.1.2
PyJWT==2.4.0
